requirements.txt
openai>=1.47
langchain-openai>=0.2.0
mcp-use>=1.3.7
trafilatura>=2.0.0,<2.1
feedparser>=6.0.11
pydantic>=2.8
python-dateutil>=2.9
fastapi>=0.112
uvicorn>=0.30
sqlite-utils>=3.36
PyYAML>=6.0
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
python-dotenv>=1.0.0
